% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeElOverblik.R
\name{dataframeElOverblik}
\alias{dataframeElOverblik}
\title{getTimeSeriesData}
\usage{
dataframeElOverblik()
}
\value{
dataframe Function returns dataframe, the DF doen contains but ensures at al modules uses the same FD
}
\description{
This function retrieves timesere data from ElOverblik
}
\details{
This set of functions is handling Token
}
