% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichRawObservations.R
\name{enrichRawObservations}
\alias{enrichRawObservations}
\title{enrichRawObservations}
\usage{
enrichRawObservations(rawObservations)
}
\arguments{
\item{rawObservations}{Observations directly from ElOverblik}
}
\value{
dataframe Returns dataframe with observations
}
\description{
This functionenric the observations from eloverblik
}
\details{
This set of functions is handling Token
}
