% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPeriodObservations.R
\name{extractPeriodObservations}
\alias{extractPeriodObservations}
\title{getTimeSeriesData}
\usage{
extractPeriodObservations(meter, period)
}
\arguments{
\item{meter}{This is your meter number identifying your meter e.g. "123456789012356789"}

\item{period}{This is the period fetched from getTimeSerie}
}
\value{
dataframe Returns dataframe with observations
}
\description{
This function retrieves timesere data from ElOverblik
}
\details{
This set of functions is handling Token
}
