% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPeriodObservations.R
\name{extractPeriodPbservations}
\alias{extractPeriodPbservations}
\title{getTimeSeriesData}
\usage{
extractPeriodPbservations(meter, period)
}
\description{
This function retrieves timesere data from ElOverblik
}
\details{
This set of functions is handling Token
}
