% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCharges.R
\name{getCharges}
\alias{getCharges}
\title{getCharges}
\usage{
getCharges(meteringPoint, data_access_token)
}
\arguments{
\item{data_access_token}{Data Access Token created by getDataAccessToken}

\item{meter}{This is your meter number identifying your meter e.g. "123456789012356789"}
}
\value{
pricing_data List of pricing data
}
\description{
This function return pricing data
}
\details{
This set of functions is handling Token
}
