% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeasuringQuality.R
\name{getMeasuringQuality}
\alias{getMeasuringQuality}
\title{getMeasuringQuality}
\usage{
getMeasuringQuality(code, lang = "UK")
}
\arguments{
\item{lang}{The language you want returned e.g. DK or UK, default = UK}

\item{core}{The quality code from the metering observation}
}
\value{
text Translated quality code
}
\description{
This functionen can only take one variable, not vector or dataframe, use lapply
}
\details{
This function translate from the quality code A1 .. A5 to text
Code = A01 DK translation = Korrigeret UK translation = Adjusted Tekst =  Korrigeret
}
