% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeterData.R
\name{getMeteringPointInfo}
\alias{getMeteringPointInfo}
\title{Get Metering Point Information}
\usage{
getMeteringPointInfo(data_access_token)
}
\value{
JSON
}
\description{
This function is not finished yet, but should get all the meering point that i have acess to
I haven testet that correct data is returned when retrieving multiple meteringPointID, since I have only access to one
}
