% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSerieData.R
\name{getTimeSeriesData}
\alias{getTimeSeriesData}
\title{getTimeSeriesData}
\usage{
getTimeSeriesData(
  meteringPoint,
  data_access_token,
  dateFrom,
  dateTo,
  aggregation = "Actual"
)
}
\arguments{
\item{meteringPoint}{The number you have recieved from you r provider}

\item{data_access_token}{Data Access Token created by getDataAccessToken}

\item{dateFrom}{start data from where we retrieve data}

\item{dateTo}{end date from which we retrieve data, note dateFrom - dateTo should not be greater than 720 days}

\item{aggregation}{see manual}
}
\value{
data_access_token This is the token you have to use for each call to andelenergi API
}
\description{
This function retrieves timesere data from ElOverblik
}
\details{
This set of functions is handling Token
}
