% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declareVariables.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Get Metering Point Information}
\format{
An object of class \code{environment} of length 1.
}
\usage{
pkg.env
}
\description{
This function is not finished yet, but should get all the meering point that i have acess to
I haven testet that correct data is returned when retrieving multiple meteringPointID, since I have only access to one
}
\details{
Global variables
}
\keyword{datasets}
