% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeElOverblik.R
\name{dataframeElOverblik}
\alias{dataframeElOverblik}
\title{dataframeElOverblik}
\usage{
dataframeElOverblik()
}
\value{
dataframe Function returns dataframe definition
}
\description{
This function retrieves timesere data from ElOverblik
}
\details{
This function defines a dataframe used by the package
}
