% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCharges.R
\name{getCharges}
\alias{getCharges}
\title{getCharges}
\usage{
getCharges(meteringPoint, data_access_token)
}
\description{
This function return pricing data
}
\details{
This set of functions is handling Token
}
