% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeteringPointInfo.R
\name{getMeteringPointInfo}
\alias{getMeteringPointInfo}
\title{Get Metering Point Information}
\usage{
getMeteringPointInfo(data_access_token)
}
\arguments{
\item{data_access_token}{The data_access_token i provided by get_data_access_token(refreshtoken)}
}
\value{
list A list containing all the basic information for the metering point
}
\description{
The purpose of this function is to extract the information related to the Metering Point.
this extraction provides the following attributes.
\itemize{
\item [streetCode] Street code – part of metering point location address.
\item [streetName] Street name – part of metering point location address.
\item [buildingNumber] Building number – part of metering point location address.
\item [floorId] Floor id – part of metering point location address.
\item [roomId] Room id – part of metering point location address.postcode String
\item [municipalityCode]
\item [locationDescription]
\item [settlementMethod]
\item [firstConsumerPartyName]
\item [secondConsumerPartyName]
\item [settlementMethod]
\item [meterReadingOccurrence]
\item [meterNumber]
\item [consumerStartDate]
}
}
\details{
The purpose of this function is to extract the information related to the Metering Point
}
