% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declareVariables.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Global variables}
\format{
An object of class \code{environment} of length 2.
}
\usage{
pkg.env
}
\description{
Global variables
}
\keyword{datasets}
